Updated microphone styling interactivity
Since the microphone is the only really used button we have to use it in innovative ways to make sure that the user understands what is happening. For this reason, after recording, the button signifies when the answer is being created. Then afterwards it also signifies the first 5 seconds of the answer being said by the avatar. After 5 seconds you can interrupt with your next question.